+++ app/src/main/java/com/github/andreyasadchy/xtra/model/misc/StvChannelResponse.kt
@@ -6,5 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 class StvChannelResponse(
     @SerialName("emote_set")
-    val emoteSet: StvGlobalResponse,
+    val emoteSet: StvGlobalResponse? = null,
+    @SerialName("emote_set_id")
+    val emoteSetId: String? = null,
 )